Enable threads for tk 8.5
We're working on packaging the [OOMMF software](http://math.nist.gov/oommf/), which benefits from threading in tcl/tk. Threads are apparently enabled by default in tk 8.6, but not 8.5. However, Python is built with tk 8.5, so we want to build OOMMF against that as well, so that it can be installed in a conda environment with Python.